+++ src/app/connection/mod.rs
@@ -1,12 +1,15 @@
 use std::{
     collections::HashMap,
     sync::mpsc::{Receiver, Sender},
+    vec,
 };
 
 use async_std::task;
 
+use kv::Key;
 use sqlx::{
-    postgres::PgConnectOptions, Column, Connection as SqlxConnection, Executor, PgConnection, Row,
+    postgres::{PgArguments, PgConnectOptions},
+    Arguments, Column, Connection as SqlxConnection, Executor, PgConnection, Row,
 };
 use uuid::Uuid;
 
@@ -64,16 +67,36 @@ impl Connection {
     pub(crate) async fn list_schemas(&mut self) -> Result<Vec<String>, sqlx::Error> {
         match &mut self.internal {
             TypedConnection::Postgres(connection) => {
-                let databases = connection
+                let schemas = connection
                     .fetch_all(sqlx::query(
                         "SELECT schema_name FROM information_schema.schemata;",
                     ))
                     .await?;
-                let schemas: Vec<String> = databases
+                let schemas: Vec<String> = schemas
                     .iter()
                     .map(|row| row.get::<String, usize>(0))
                     .collect();
                 Ok(schemas)
+            }
+        }
+    }
+
+    pub(crate) async fn list_tables(&mut self, schema: &str) -> Result<Vec<String>, sqlx::Error> {
+        match &mut self.internal {
+            TypedConnection::Postgres(connection) => {
+                let mut args = PgArguments::default();
+                args.add(schema);
+                let tables = connection
+                    .fetch_all(sqlx::query_with(
+                        "SELECT table_name FROM information_schema.tables where table_schema = $1;",
+                        args,
+                    ))
+                    .await?;
+                let tables: Vec<String> = tables
+                    .iter()
+                    .map(|row| row.get::<String, usize>(0))
+                    .collect();
+                Ok(tables)
             }
         }
     }
@@ -90,13 +113,29 @@ pub(crate) struct ConnectionsManager {
     pub(crate) connections: HashMap<ConnectionKey, Connection>,
 }
 
+#[derive(PartialEq, PartialOrd, Clone, Eq, Debug)]
 pub(crate) enum DbRequest {
-    ListSchemas { server_id: Uuid, database: String },
-
+    ListSchemas {
+        server_id: Uuid,
+        database: String,
+        retries: i32,
+    },
+    ListTables {
+        server_id: Uuid,
+        database: String,
+        schema: String,
+        retries: i32,
+    },
     ListDatabases(Uuid),
     ConnectToServer(Uuid, String),
     ConnectToDatabase(Uuid, String, String),
-    Execute(Uuid, String, String),
+    Execute(Uuid, String, String, i32),
+}
+
+#[derive(PartialEq, PartialOrd, Clone, Eq, Debug)]
+pub(crate) enum DownConnectionReason {
+    IoError(String),
+    MissingConnection,
 }
 
 #[derive(PartialEq, PartialOrd, Clone, Eq, Debug)]
@@ -107,10 +146,20 @@ pub(crate) enum DbResponse {
         database: String,
         schemas: Vec<String>,
     },
+    TablesListed {
+        server_id: Uuid,
+        database: String,
+        schema: String,
+        tables: Vec<String>,
+    },
     Connected(Uuid),
     Executed(Uuid, Vec<String>, Vec<Vec<String>>),
     Error(Uuid, String),
-    ConnectionIsDown(Uuid, String, String),
+    // ConnectionIsDown(Uuid, String, String),
+    ConnectionIsDown {
+        original_request: DbRequest,
+        reason: DownConnectionReason,
+    },
     None,
 }
 
@@ -122,11 +171,28 @@ impl ConnectionsManager {
         }
     }
 
-    fn process_db_error(error: sqlx::Error, id: Uuid) -> DbResponse {
+    fn process_db_error(
+        &mut self,
+        key: &ConnectionKey,
+        error: sqlx::Error,
+        id: Uuid,
+        repeat: Option<&dyn Fn() -> DbRequest>,
+    ) -> DbResponse {
         tracing::error!("Error executing query: {:?}", error);
         match error {
             sqlx::Error::Database(db_error) => {
                 return DbResponse::Error(id, db_error.message().to_string())
+            }
+            sqlx::Error::Io(io_error) if repeat.is_some() => {
+                if let Some(repeat) = repeat {
+                    self.connections.remove(&key);
+                    DbResponse::ConnectionIsDown {
+                        original_request: repeat(),
+                        reason: DownConnectionReason::IoError(format!("IO Error: {:?}", io_error)),
+                    }
+                } else {
+                    DbResponse::Error(id, format!("IO Error: {:?}", io_error))
+                }
             }
             _ => DbResponse::Error(id, format!("unknown db error: {:?}", error)),
         }
@@ -134,74 +200,128 @@ impl ConnectionsManager {
 
     fn process_request(&mut self, request: DbRequest) -> DbResponse {
         match request {
-            DbRequest::ListSchemas {
+            DbRequest::ListTables {
                 server_id,
                 database,
+                schema,
+                retries,
             } => {
-                let key = ConnectionKey {
+                let connection_key = ConnectionKey {
                     name: database.to_string(),
                     server_id,
                 };
-                if let Some(connection) = self.connections.get_mut(&key) {
+                let repeat = || DbRequest::ListTables {
+                    server_id,
+                    database: (&database).to_string(),
+                    schema: (&schema).to_string(),
+                    retries: retries + 1,
+                };
+                if let Some(connection) = self.connections.get_mut(&connection_key) {
+                    match task::block_on(connection.list_tables(&schema)) {
+                        Ok(tables) => DbResponse::TablesListed {
+                            server_id,
+                            database: database.to_string(),
+                            schema: schema.to_string(),
+                            tables,
+                        },
+                        Err(e) => {
+                            self.process_db_error(&connection_key, e, server_id, Some(&repeat))
+                        }
+                    }
+                } else {
+                    DbResponse::Error(server_id, "No connection to database".to_string())
+                }
+            }
+            DbRequest::ListSchemas {
+                server_id,
+                database,
+                retries,
+            } => {
+                let connection_key = ConnectionKey {
+                    name: database.to_string(),
+                    server_id,
+                };
+                let repeat = || DbRequest::ListSchemas {
+                    server_id,
+                    database: database.to_string(),
+                    retries: retries + 1,
+                };
+                if let Some(connection) = self.connections.get_mut(&connection_key) {
                     match task::block_on(connection.list_schemas()) {
                         Ok(tables) => DbResponse::SchemasListed {
                             server_id,
                             database: database.to_string(),
                             schemas: tables,
                         },
-                        Err(e) => Self::process_db_error(e, server_id),
+                        Err(e) => {
+                            self.process_db_error(&connection_key, e, server_id, Some(&repeat))
+                        }
                     }
                 } else {
-                    DbResponse::Error(server_id, "No connection to database".to_string())
+                    DbResponse::ConnectionIsDown {
+                        original_request: repeat(),
+                        reason: DownConnectionReason::MissingConnection,
+                    }
                 }
             }
             DbRequest::ConnectToServer(id, url) => {
-                let key = ConnectionKey {
+                let connection_key = ConnectionKey {
                     name: DEFAULT_MANAGEMENT_DATABASE.to_string(),
                     server_id: id,
                 };
-                if self.connections.contains_key(&key) {
+                if self.connections.contains_key(&connection_key) {
                     return DbResponse::Connected(id);
                 }
-                match task::block_on(Connection::connect(&key.name, &url)) {
+                match task::block_on(Connection::connect(&connection_key.name, &url)) {
                     Ok(connection) => {
-                        self.connections.insert(key, connection);
+                        self.connections.insert(connection_key, connection);
                         DbResponse::Connected(id)
                     }
-                    Err(e) => Self::process_db_error(e, id),
+                    Err(e) => self.process_db_error(&connection_key, e, id, None),
                 }
             }
             DbRequest::ConnectToDatabase(id, name, url) => {
-                let key = ConnectionKey {
+                let connection_key = ConnectionKey {
                     name,
                     server_id: id,
                 };
-                if self.connections.contains_key(&key) {
+                if self.connections.contains_key(&connection_key) {
                     return DbResponse::None;
                 }
-                match task::block_on(Connection::connect(&key.name, &url)) {
+                match task::block_on(Connection::connect(&connection_key.name, &url)) {
                     Ok(connection) => {
-                        self.connections.insert(key, connection);
+                        self.connections.insert(connection_key, connection);
                         DbResponse::None
                     }
-                    Err(e) => Self::process_db_error(e, id),
+                    Err(e) => self.process_db_error(&connection_key, e, id, None),
                 }
             }
-            DbRequest::Execute(id, name, query) => {
-                match task::block_on(self.execute(&query, id, &name)) {
+            DbRequest::Execute(id, name, query, retries) => {
+                let connection_key = ConnectionKey {
+                    name: name.to_string(),
+                    server_id: id,
+                };
+                let repeat =
+                    || DbRequest::Execute(id, (&name).clone(), (&query).clone(), retries + 1);
+                match task::block_on(self.execute(&query, &connection_key)) {
                     Ok(Some((headers, data))) => DbResponse::Executed(id, headers, data),
-                    Ok(None) => DbResponse::ConnectionIsDown(id, name, query),
-                    Err(e) => Self::process_db_error(e, id),
+                    Ok(None) => DbResponse::ConnectionIsDown {
+                        original_request: repeat(),
+                        reason: DownConnectionReason::MissingConnection,
+                    },
+                    // Ok(None) => DbResponse::ConnectionIsDown(id, name, query),
+                    Err(e) => self.process_db_error(&connection_key, e, id, Some(&repeat)),
                 }
             }
             DbRequest::ListDatabases(id) => {
-                if let Some(connection) = self.connections.get_mut(&ConnectionKey {
+                let connection_key = ConnectionKey {
                     name: DEFAULT_MANAGEMENT_DATABASE.to_string(),
                     server_id: id,
-                }) {
+                };
+                if let Some(connection) = self.connections.get_mut(&connection_key) {
                     match task::block_on(connection.list_databases()) {
                         Ok(databases) => DbResponse::DatabasesListed(id, databases),
-                        Err(e) => Self::process_db_error(e, id),
+                        Err(e) => self.process_db_error(&connection_key, e, id, None),
                     }
                 } else {
                     DbResponse::Error(id, "No connection to management database".to_string())
@@ -229,14 +349,9 @@ impl ConnectionsManager {
     async fn execute(
         &mut self,
         query: &str,
-        id: Uuid,
-        name: &str,
+        key: &ConnectionKey,
     ) -> Result<Option<(Vec<String>, Vec<Vec<String>>)>, sqlx::Error> {
         tracing::info!("Executing query: {}", query);
-        let key = ConnectionKey {
-            name: name.to_string(),
-            server_id: id,
-        };
         let connection = match self.connections.get_mut(&key) {
             Some(connection) => connection,
             None => return Ok(None),