+++ src/components/browser.rs
@@ -16,6 +16,7 @@ pub enum BrowserTreeId {
     Server(Uuid),
     Database(Uuid, String),
     Schema(Uuid, String, String),
+    Table(Uuid, String, String, String),
 }
 
 impl BrowserTreeId {
@@ -56,6 +57,18 @@ impl BrowserTreeId {
                     parts.next()?.to_string(),
                 ))
             }
+            "table" => {
+                let node: &str = parts.next()?;
+                let mut parts = node.splitn(4, ':');
+                let server_id = parts.next()?;
+                let server_id = Self::parse_server_uuid(server_id)?;
+                Some(Self::Table(
+                    server_id,
+                    parts.next()?.to_string(),
+                    parts.next()?.to_string(),
+                    parts.next()?.to_string(),
+                ))
+            }
             _ => None,
         }
     }
@@ -66,6 +79,9 @@ impl BrowserTreeId {
             Self::Database(server_id, name) => format!("database:{}:{}", server_id, name),
             Self::Schema(server_id, database, name) => {
                 format!("schema:{}:{}:{}", server_id, database, name)
+            }
+            Self::Table(server_id, database, schema, table) => {
+                format!("table:{}:{}:{}:{}", server_id, database, schema, table)
             }
         }
     }
@@ -174,6 +190,49 @@ impl Component<Msg, TisqEvent> for BrowserTree {
                 return Some(Msg::None);
             }
 
+            Event::User(TisqEvent::DbResponse(DbResponse::TablesListed {
+                server_id,
+                database,
+                schema,
+                tables,
+            })) => {
+                // tracing::debug!("Schemas listed: {:?}", tables);
+                let schema_id = BrowserTreeId::Schema(server_id, database.clone(), schema.clone());
+                let tree_id = schema_id.to_string();
+                {
+                    let tree = self.component.tree_mut();
+                    let node = tree.root_mut().query_mut(&tree_id)?;
+                    node.clear();
+                    for table in tables {
+                        let id = BrowserTreeId::Table(
+                            server_id,
+                            database.clone(),
+                            schema.clone(),
+                            table.clone(),
+                        )
+                        .to_string();
+                        let mut table_node = Node::new(id, table.clone());
+
+                        // dummy is created to make the table node expandable
+                        // it is a workaround for the limitation of the treeview component
+                        // let dummy: Node = Node::new(
+                        //     format!("dummy:{}/{}/{}/{}", server_id, database, schema),
+                        //     "loading tables...".to_string(),
+                        // );
+                        // table_node.add_child(dummy);
+
+                        node.add_child(table_node);
+                    }
+                }
+                // This solution can work as well without blocking, but requires
+                // open access to open_node method as well as new one tree_state_mut
+                let node = self.component.tree().root().query(&tree_id)?.clone();
+                let root = self.component.tree().root().clone();
+                self.component.tree_state_mut().open_node(&root, &node);
+
+                return Some(Msg::None);
+            }
+
             Event::User(TisqEvent::DbResponse(DbResponse::SchemasListed {
                 server_id,
                 database,
@@ -190,13 +249,13 @@ impl Component<Msg, TisqEvent> for BrowserTree {
                             .to_string();
                         let mut schema_node = Node::new(id, schema.clone());
 
-                        // dummy is created to make the server node expandable
+                        // dummy is created to make the schema node expandable
                         // it is a workaround for the limitation of the treeview component
-                        // let dummy: Node = Node::new(
-                        //     format!("dummy:{}/{}/{}", server_id, database, table),
-                        //     "loading columns...".to_string(),
-                        // );
-                        // table_node.add_child(dummy);
+                        let dummy: Node = Node::new(
+                            format!("dummy:{}/{}/{}", server_id, database, schema),
+                            "loading tables...".to_string(),
+                        );
+                        schema_node.add_child(dummy);
 
                         node.add_child(schema_node);
                     }
@@ -303,7 +362,15 @@ impl Component<Msg, TisqEvent> for BrowserTree {
                             return Some(Msg::OpenConnection(server_id))
                         }
                         Some(BrowserTreeId::Database(server_id, database)) => {
-                            return Some(Msg::OpenDatabase(server_id, database))
+                            return Some(Msg::OpenDatabase(server_id, database, 0))
+                        }
+                        Some(BrowserTreeId::Schema(server_id, database, schema)) => {
+                            return Some(Msg::OpenSchema {
+                                server_id,
+                                database,
+                                schema,
+                                retries: 0,
+                            })
                         }
                         _ => self.perform(Cmd::Custom(TREE_CMD_OPEN)),
                     },